cdc/requirements.txt
requests==2.31.0
python-decouple==3.8
psycopg2==2.9.6